rewriting functions to allow for an extended body message to be displayed when clicking "see more". reorganizing files and moving functions to separate the map functions. created additional examples in api to test the desired results are working properly
