+++ client/src/components/reviews/apiExample.js
@@ -10,7 +10,7 @@ let data = {
         "summary": "ReviewTitle",
         "recommend": true,
         "response": null,
-        "body": "Review Body For The Review That We Are Testing This Review Text Must Meet A Minumum Number Of Characters To Work But I Dont Know Why I Did A Camel Case Sentence Thats Just Weird",
+        "body": "Review Body For The Review That We Are Testing This Review Text Must Meet A Minumum Number Of Characters To Work But I Dont Know Why I Did A Camel Case Sentence Thats Just Weird. Why can I simply look out the window for hours on end. I love my life, I love the weather, I love nature. Please let this sentence be long enough now. I do not care about this product really. Shhhhhhh, it is a secret!",
         "date": "2022-01-04T00:00:00.000Z",
         "reviewer_name": "ReviewerName",
         "helpfulness": 0,
