+++ client/src/components/reviews/reviewListComponents/ReviewsList.jsx
@@ -19,24 +19,24 @@ const ReviewsList = ( {productID} ) => {
   let [wholeReviewList, setWholeReviewList] = useState(reviewArray)
 
   // on initial render, GET data for productID
-  useEffect(()=>{
-    axios({
-      method: 'get',
-      url: `https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfe/reviews?product_id=${productID}&page=1&count=50`,
-      headers: {
-        'Authorization': API_KEY
-      }
-    })
-      .then(({data})=>{
-        // console.log('I am the FIRST request for data', data);
-        reviewArray = data.results;
-        setWholeReviewList(data.results);
-        setCurrentDisplay(reviewArray.slice(0, reviewIndex));
-      })
-      .catch((err) => {
-        console.log(err);
-      })
-  }, []);
+  // useEffect(()=>{
+  //   axios({
+  //     method: 'get',
+  //     url: `https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfe/reviews?product_id=${productID}&page=1&count=50`,
+  //     headers: {
+  //       'Authorization': API_KEY
+  //     }
+  //   })
+  //     .then(({data})=>{
+  //       // console.log('I am the FIRST request for data', data);
+  //       reviewArray = data.results;
+  //       setWholeReviewList(data.results);
+  //       setCurrentDisplay(reviewArray.slice(0, reviewIndex));
+  //     })
+  //     .catch((err) => {
+  //       console.log(err);
+  //     })
+  // }, []);
 
   let [currentDisplay, setCurrentDisplay] = useState(reviewListExample);
   let [reviewIndex, setReviewIndex] = useState(2);
@@ -48,24 +48,24 @@ const ReviewsList = ( {productID} ) => {
     setCurrentDisplay(wholeReviewList.slice(0, reviewIndex))
   }, [reviewIndex]);
 
-  useEffect(()=>{
-    // when the sorting method changes, make new GET request to get data sorted appropriately
-    axios({
-      method: 'get',
-      url: `https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfe/reviews?product_id=${productID}&page=1&count=50&sort=${sortFilter}`,
-      headers: {
-        'Authorization': API_KEY
-      }
-    })
-      .then(({data})=>{
-        reviewArray = data.results;
-        setWholeReviewList(data.results);
-        setCurrentDisplay(reviewArray.slice(0, reviewIndex));
-      })
-      .catch((err) => {
-        console.log(err);
-      })
-  }, [sortFilter])
+  // useEffect(()=>{
+  //   // when the sorting method changes, make new GET request to get data sorted appropriately
+  //   axios({
+  //     method: 'get',
+  //     url: `https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfe/reviews?product_id=${productID}&page=1&count=50&sort=${sortFilter}`,
+  //     headers: {
+  //       'Authorization': API_KEY
+  //     }
+  //   })
+  //     .then(({data})=>{
+  //       reviewArray = data.results;
+  //       setWholeReviewList(data.results);
+  //       setCurrentDisplay(reviewArray.slice(0, reviewIndex));
+  //     })
+  //     .catch((err) => {
+  //       console.log(err);
+  //     })
+  // }, [sortFilter])
 
 
   return (<ReviewsContainer data-testid="reviewListDivContainer">
